FIX :: 안넘어가는 오류 수정
